Add authentication test code
Why
An authentication flow needs to be set-up so this app can access the
user data.

How
* Add the basic authentication flow instructions from the Spotipy website.
It's not what I'll end up using, but it's good to have a record of what works.

Note
This requires user input each time to authorize the transaction so I need
to fix that so it's more automated.