Add recently played api call
Why
The main goal is to use the recently played tracks to make predictions.
In sound.py, that scope and call should be used.

How
* Add the necessary scope and API call for getting recently played tracks

Note
This is still in the testing phase since I need to determine the time frame
to use. I also need to figure out how to set-up automatic authentication so
I can run it nightly.