Fix issues with json output
Why
The json output was full of unnecessary special characters (ex. newline)
which made it hard to read and difficult to query through SQL.

How
* Remove the json_res variable and the json.dumps. This was causing part
of the formatting issue.
* Add encoding = 'utf-8' to the test file open code since this prevents
errors with some of the characters from the API pull.
* Add the results variable to the json.dump function. This is all that's needed
to load the json data.

Note
I also removed some of the logic for returning the list of top songs since
it's irrelevant.