First draft of track metrics functions
Why
I'm combining the functions for track features and track analysis into
one file since the functions are relatively small and they accomplish a
similar goal of providing detailed information on each track.

How
* Remove the separate track_analysis and track_features files and replace them
with track_metrics.
* Import the necessary libraries that will make the functions work
* Use spotify client credentials for authorization since no user authentication
is required.
* Add a variable that stores the list of track IDs since this will be used in the
subsequent functions.
* Create the features function to get the track features data. This is a simple
call of the audio_features API. It can accept a list of IDs as a parameter so there's
no need for additional logic.
* Create the analysis function for getting the audio analysis data. This is more
complicated since only 1 track ID is allowed per call. As a result, I used a for loop
to iterate through each track ID in the track list, get the json object for that ID, and
add it to a list of all the audio analyses for the tracks.

Note
The audio_analysis function is compute intensive and also doesn't seem to provide
data that is as useful as the track features function call. However, I'm going to
include it anyway in case it's useful in the future.