Add initial API logic (#1)
Why
This app will use the Spotify web API to produce a prediction of a mood state based on their daily music habits. The hypothesis is that music can be used to classify a person's mood.

How

- Import the Spotipy module to pull data on the recently played tracks and the features for those tracks.
- Add functions for producing the connection info needed for the API requests
- Add functions to pull the recently played track IDs and their features.

Note
The next step would be to start loading these files into a postgres database using python.