WIP: Code for loading the json data
Why
There needs to be a way to take the json provided in track_info and
track_metrics and load it into the postgres database.

This is still a work in progress and as of this commit doesn't work.

How
- Import the necessary libraries. Make sure to use the sys.path.append
to gain access to the functions in the data_pull folder
- Establish the variables that will be used in the SQL statements.
- Pull in the recently_played and features data
- Open a cursor so the SQL statements can be executed.
- Add a top level try/finally that closes the connection and cursor once
the SQL statements finish or fail.
- Add a for loop that goes through each table in the table list and executes
a create table and insert statement against that table.
- If the above fails, throw a programming error with a message saying that
the insert failed.