WIP: Clean up the table load code
Why
I still don't hae it in the spot it needs to be, but cleaning up the variables
and the queries helped prevent some syntax errors and made the code more readable.

How
- Remove the db, schema, column, and column_type variables. They are unnecessary
since they can just be written within the SQL code. This also prevents the
syntax errors I was getting when trying to parameterize these variables.
- Clean up the for loop so the table create script doesn't parameterize any
of the fields. This might make the table vulnerable to SQL injection, but I'm
not sure.
- Only use the if condition on the insert clause and only parameterize the
actual value. This should be sufficient to prevent sql inject.

Note
It's still not perfect and for some reason I couldn't get any of the API code
to run so there may be an issue with the Spotify API. I'll check on it again
later.