Fix the load_tables script
Why
The script wasn't working so there were several changes that needed to be added

How
- Remove unnecessary libraries
- Add the datetime and logging libraries since those are needed for
logging the execution status
- Comment out the track function imports since those still need to be fixed
- Add additional variables for the raw db and the spotify schema
- Add temporary info_data and feature_data string variables to test the insert
- Add a dictionary for the table_data. This way, any new tables can be easily added to
the dictionary and looped through without having to add an if/else statement
to check the table to load
- Add logging config.
- Add code under the top level try block that says that the job is initiated. This
includes a timestamp field to help with tracking.
- Use with cursor before the subset of try blocks. This way it can be used across
all the sql executions. The with makes sure to close it once it's complete.
- Add logic to create the schema if it doesn't exist. A try/except/else
is used with logging to make sure common errors are handled and tracked.
- Change the for loop to iterate through the dictionary so both the table (key) and
the data (value) can be used.
- Remove the if/elif code since the dictionary approach makes it obsolete
- Add similar try/except and logging logic for the create table and insert
table logic.
- Add logging to the finally block since it's good to know that the data load
is complete and the connection closed.