Add test json data to variables
Why
Eventually I'll be loading the spotify data into these tracks tables.
Before I do that I need to make sure that it's possible to load data in
this format to the database. I also made some changes to error catching and
logging to make the logs more readable

How
- Hardcode json style data to the info_data and feature_data variables
- For all else code in the try blocks, use logging.info instead of logging.exception
- Make the create table and insert data variables multi line to improve readability
- Update the create table string to have  id and json columns
- Update the insert string to include the src column specifically.