Load the spotify api data
Why
Now that the load tables script is able to load static data, it's time to adjust
it so it can load the dynamic api data from spotify.

How
- Uncomment the import for the tracks modules
- Assign the results of recently_played and track_features to variables
- Replace single quotes in those variables with '' since that's the simplest way to
upload single quotes to postgres.
- Add the created_on timestamp field to the tables. This is necessary for determining
which records to use since we only want to use the most recent api pull for predictions.
- Use utcnow as the value for the timestamp in the insert code