Test out token authentication
Why
In order to automate the postgres pull, I need a way to keep the token
refreshed so it's not required to log-in between sessions. This commit
represents the initial attempt at identifying the token.

How
- Clean-up the oauth function. There were some unnecessary
pieces of code that could be removed
- Add the initial token info to the oauth function so that can get returned too

Note
This isn't working yet, but it's a start and should be committed.