Add logic to refresh the access token
Why
Currently, if the access token expires, the user is prompted to log in
again. This needs to be changed so that the token is refreshed if it expires.

How
All of these changes take place in the oauth connection function
- Remove the oauth_conn variable. It's not necessary anymore since the
oauth connection will be using the token
- Add an if clause that checks if there's token_info and assigns the access
token to the token variable. If no token info exists, the user is prompted
to log in so an access token can be set to the token variable
- Next, an if clause is used to check if the token is expired. If it is,
it runs the refresh access token function and sets the auth_token variable
equal to the auth_token connection logic. Otherwise, it uses the existing
sp connection logic as the auth_token