Add exception handling and logging
Why
Since an API is used, tracks.py is vulnerable to issues when trying to
gather the data. As a result, good exception handling and logs are critical
to the program working.

How
- Remove the logging config from load_tables and add it to tracks. Only
one config is needed.
- Rename the log file to execute since it encompasses more than just
the tracks data
- Add try/except blocks to the recently played function for both the API
call and the conversion to JSON
- Add a try/except block to the track features function for both the API
call and the conversion to JSON.

Note
I had to use the generic Exception for the API calls. It would be better
to use specific exceptions such as those provided by Spotipy, but for now
this is suitable for catching any errors that I might not be aware of.