Add messaging capabilities
Why
I want to know if an API job succeeds or fails. This can be accomplished
by sending texts depending on the status of the job.

How
transmit.py
- Copy the code from the publish.py script in Nuclear Octopus and make
adjustments so it sends the proper message
- Also, add a parameter to the communicado function so it uses the success
value produced in the load_tables script

load_tables.py
- Add a success value to the try blocks that inserts the data. If it succeeds,
then the text says that. Otherwise, if an exception is registered the success
value is set to false

execute_load.py
- Pull in the communicado function and use that to send the text based on
the results from the load_tables.py extract_load function.