Add try/except logic to connections code
Why
It's possible that the program will fail due to connection related issues.
There should be logic that catches and logs these issues.

How
- Import logging so that can be used for logging the messages
- Import urllib3 and the necessary exceptions
- Add try/except logic to connections.py so any HTTP based exceptions such
as NewConnectionError are caught and logged.

Note
-I also made some minor spacing changes to load_tables.py that are included here.

Set default success parameter

Why
Even though the success value should always end up being either true or false
when it's fed to the communicado function, a default value is a good safety net.

How
- In the communicado function, set the default success value to False

Add spotify error exception

Why
The SpotifyOauthError exception shows up when the connection to the oauth
flow fails. This is a good one to catch in the connections logic.

How
- Add the SpotifyOauthError exception to the list of exceptions used when
the oauth connection function raises and exception.

Note
I also included the shebang from the linux version of execute_load since it
is helpful to have it there and it doesn't seem to break anything on windows.

Add responses data pull file

Why
This file will pull data from the google forms responses and create a csv
file that can be used to produce a dataframe. Most of the code is the same
as what's in Nuclear Octopus.

How
- Import libraries needed for the code
- Set-up authorization credentials for accessing the Google Sheets files
that store the response data
- Create the fetch_data function to pull the responses and create a csv file
- In the try/except block for the csv creation, make sure IOErrors are caught
and logged. Successes should be logged as well
- It's import to return the target_csv variable since it is used in the
response prep file to call the function.

Add responses_prep file

Why
This file will be used to create a dataframe that can be uploaded to postgres
in the load_tables script. It is similar to what's in Nuclear Octopus, but it
removes most transformation logic since that can all be done in SQL. This will
make the python program run more efficiently.

How
- Load the necessary libraries
- Create the data_prep function that pulls in the csv file with the responses
and returns a dataframe for the postgres load that only contains the index,
the timestamp, and the label (dependent variable).

Rename load_tables file

Why
This file should only be used for loading tracks. Since responses data
from Google Sheets will also be used, a new name is required.

How
- Rename the file to load_tracks

Minor changes to a few files

Why
Just committing a few minor un-related changes since it doesn't make
sense to have a spearate commit message for each

How
- Delete responses_prep.py since it is obsolete.
- Remove whitespace in responses.py
- Fix comments on tracks.py

Create file to create a responses json object

Why
The responses data from the original Nuclear Octopus app is valuable and
is worth keeping in the postgres database. The best way to store it will
be through json files.

How
- Bring in the libraries needed.
- Set-up logging
- Create a function called get_responses which will pull the csv data
into a dataframe and then parse it into json
- Set-up a try block that creates a dataframe from the csv and throws an
error if it can't
- use "replace" to replace all NaN values with null
- For all datetime fields, set the value to null if it's NaT and convert everything
else to string.
- Create a list of dictionaries out of the response values that will be used to convert the data
- Create a metadata dictionary which will be added to each dictionary object.
This is useful for auditing the data and preventing duplicates.
- Use json dumps to create a json object that can e loaded into postgres.

Add logging to each necessary file

Why
I'm not sure if the logging config needs to be added to each file, but so
far it looks like not everything is being logged so I added the logging config
to every file that uses logging.

Rename responses file that produces json

I renamed the file from responses_json to responses_prep since the variable
returned in the function is also responses_json.

Rename load tables to load tracks

Why
Since load_tables is now called load_tracks, the change needs to be reflected
in the other files that rely on it.

How
- Rename the extract_load function in load_tracks to tracks_to_pg since that
makes more senses.
- Update execute_load.py so it uses the correct file and function import.

Remove dataframe connection vars

Why
Right now, it doesn't make sense to use sql alchemy to push the data to
postgres so these changes can be removed.

How
- Remove the flavor and library variables
- Remove the dataframe_conn object

WIP: Add logic to load survey responses

Why
It will be helpful to have this data inthe postgres db since it can be combined
with the API data to make predictions. This is still a work in progress
since the data is loading in a format that isn't useable. However, I wanted
to get this code included in the commit history so I can revert to it if
necessary.

How
- Set-up logging
- Rename function to responses_to_pg
- Rename the schema to surveys
- Use similar logic to what's on load_tracks where the schema is created,
the table is created, and the data is inserted.
- The main problem with this script currently is that all of the responses
get uploaded in one row. One workaround is to loop through these using a for loop,
but that's not performant and also still doesn't provide a great looking table
in postgres

Basic changes to files

Just changed the wording in some comments and moved others. Also, had
to add a pylint argument at the top of the file to deal with the false positives
from pylint against the errors module from psycopg2.

Delete old response records from db

Why
Since the historical json data will be the same for every daily load, only
the record that includes the most recent response is needed. Deleting the
old will save ono query time since it won't be necessary to find the record
with the max created date.

How
- In the try logic for inserting the json data, add another section that
uses a delete from on the table to remove all records with a created on date
before the timestamp value that was used to set the created on for the uploaded
record.