Add spotify error exception
Why
The SpotifyOauthError exception shows up when the connection to the oauth
flow fails. This is a good one to catch in the connections logic.

How
- Add the SpotifyOauthError exception to the list of exceptions used when
the oauth connection function raises and exception.

Note
I also included the shebang from the linux version of execute_load since it
is helpful to have it there and it doesn't seem to break anything on windows.