Get branch ready for raspberry pi (#9)
Why
I'll be running the scripts regularly using raspberry pi as a pseudo server. There are some breaking changes as well as nice-to-haves that can be added to the branch so it works better on the rpi environment.

How
- Add exception handling to the connection string
- Make other minor adjustments

Note
There's a shebang in the execute_load.py file which doesn't make sense for windows, but is important for linux. It doesn't harm windows so it's ok being there.