Add text message conditional logic (#10)
Why
If the data load is a success, it's not necessary to send a text message. A message should only be sent if the load fails

How
- Add if/else logic so if "success" is true, nothing is done. Otherwise, a text message will be sent because it means there was an issue with the load.
- Wrap the else in a try/finally block so the connection will be closed even if something happens in the try block.

Note
I could move the if/else logic further up in the code since most of the logic is unnecessary if the success value is true. However, I might want to end up sending messages from time to time even if success is true so I left it this way.