Add conditional logic to text messaging
Why
Since there are two separate jobs to load the data, both will return
a success variable that can be used to determine if a failure text message
should be sent.

How
- Import logging since this is useful for auditing the data
- Add an if/else block that checks if the tracks or responses loads failed.
If either failed, a failure text will be sent. Otherwise, no text is sent
and the success is logged