Remove/add fields to API dictionary
Why
Some of the *string fields such as codestring take up too much space and
aren't useful for any type of analytics at this point. Also, the track ID
needs to be added to the dictionary since that is needed for joining in the
star schema when creating the fact.

How
- Use the pop method to remove the *string fields from the nested track
dictionary
- Add the id field to the nested track dictionary and set it to equal the current
track.