Add track_analysis function to load_tracks
Why
This function will be used in a similar manner to the regular loads track
function, but since it has more nested dictionaries a separate function is
needed.

How
- Create a function called analysis_to_pg.
- Add the same logic as tracks_to_pg, but iterate through the list of tracks
 in the dictionary. Store the new object as json_dict.
- Update the insert try block to upload the json_dict.

Note
This table takes over 2 minutes to query in the database and also runs slow.
However, this data could prove to be valuable for machine learning since
it's so granular.