Add table group to function
Why
It's more efficient to iterate through each function on execute_load.py
by using a dictionary composed of a simple name for the function and the
success value (true/false.)  As a result, the communicado function in the
transmit.py script can be updated to pull in that table group to be used in
the message.

How
- Add the table group to the function arguments and as parameters in
the message sent.

Note
I also commented out the if clause that only sends a message on failure since
I want to make sure this is still working