Add track analysis to execute script.
Why
Since the track analysis is now accessible, it needs to be pulled into
postgres on the same schedule as the other sources.

How
- Import the sleep function. It will be needed for adding a delay befor the
track analysis function runs.
- Create a dictionary of the success values for each functions. This set-up
will be increasingly important as more data loads are added.
- Iterate through each key/value in the success dictionary so the communicado
function can run for each table group and success value to determine if a
message should be sent and what the contents of the message are.
-