Load audio analysis data souce (#12)
Why
Spotify provides an API that returns granular data on each track. This data could be helpful in producing predictions since it goes to the lowest level of a track's attributes.

How
- Create functions that pull the audio analysis data and load it into postgres
- Add the function call to the execute_load script so it runs on a schedule.

Note
I also made some adjustments and improvements to other pieces of the code, but those changes were not the main goal of this branch.