Ad basic comment to delete block
Just need to explain further why I'm deleting the records rather than
doing a create/replace on the table. It's because I don't want to lose the
prior record if something happens in the table create phase. Alternatively,
I could put something that reverses the prior transactions, but this
solution is fine for now.