Pull and load artists data (#19)
Why
The random forest classification will be even stronger if specific artist
attributes can be used as features.

How
- Pull in artists using the Spotipy API
- Load the artists data to the postgres db

Note
I also removed the track analysis functions since that data takes up too much storage at the moment and is not valuable enough to keep it in the current code. I can always refer back to it in prior commits if necessary.