Hide the responses function calls
Why
The survey responses data will be necessary for random forest classification.
However, attempting to pull this data on a raspberry pi results in errors
even though it works fine on PC. Until I can figure out how to fix this
issues so it runs on both platforms, I'm going to remove the responses
functions from the execute script.

How
- On execute_load.py, remove all the function calls, imports, etc. related
to the survey response code. This can be added in later once the raspberry pi
issues are fixed.