Fix sys path for file imports
Why
Files in the project that come from other directories need to have their path
set in the file where they are imported. There is a cleaner way to do this
using os.path.join

How
- For the specific path in the sys.path.append, use os.path.join to connect
each folder in the path.
- Remove the replace from user_home since os.path.join makes it unnecessary
- Add init.py files to the folders where the files are imported into files in other
directories.

Note
Eventually, it will make sense to create packages for these files, but for now
this solution works.