Fix path append and add responses function
Why
The format for sys.path.append should be updated to match what's in the
other files. Also, the responses function should be run so that data gets
populated as well. I previously excluded it since pandas wasn't working properly on
the raspberry pi, but now that I'm transitioning to heroku it will work fine.

How
- Add the os.path.join approach to the sys.path.append code.
- Add the response_to_pg() output to the success dict so it is looped through
when executing the script