Change how the app works in heroku
Why
Instead of a pointless flask app, a simple python script with the necessary
operations can be used.

How
- Replace the app.py code with the cache file creation code
- Remove the web process from the Procfile and add a worker process that
runs the new app.py script to create the cache file

Note
I'm not sure if this will actually work, but it's a more sensible approach to
the goals of this app.