Add the cache create/overwrite code
Why
Since execute_load is the file that's used to run the rest of the code,
it makes sense to create or overwrite the cache file here.

How
- Add the cache create code at the beginning of the file.

Note
It will be necessary to scale to more users eventually so a function in
a class could make sense for this.