Fix Spotify token issue (#40)
Why
The cache file for that held the refreshable Spotify token was dropped once the dyno process completed in Heroku. This meant that I had to re-authorize it manually in Spotify.

How
- Add code to execute_load.py that creates/overwrites the .cache file with the contents of the SPOTIPY_CACHE variable. This allows the program to operate the same as before.

Note
I tried a lot of approaches, but this was the simplest for my current needs. If more users are added, something more scalable will be needed. A user class with this as a function would be effective.