Add logic for deleting duplicates
Why
In order to keep the table as slim as possible, any duplicate records that
are loaded should be removed. A record is considered a duplicate if the src
text value matches the src value from a later record.

How
- Add another try/except block as an else after the data is inserted.
- UJse a CTE to find the max ID for each src value. This will grab the
latest version of each src text value, although it doesn't really matter if it's the
latest ID or the earliest.
- Use the src and max ID to identify the records that should be deleted in the
delete from.
- Add the necessary exception handling
- Also convert the json fields to jsonb since this will help json operations
in dbt