Add initial user class
Why
The user class will be needed for determining which data to pull and how
to load it into postgres. The basic user data will be stored in a table
in postgres so that will have to get pulled in.

How
- Add a user file with the "User" class that initializes a user object with
the fields needed to pull and load the data

Note
A lot of the spotify attributes are sensitive and shouldn't be stored in
the db or the code. I'll have to figure out a different way to pull them
in for each user.