Remove text messaging files
Why
Sending a text message on failure was useful for alerting me to failures.
However, now I have rollbar so it's deprecated.

How
- Remove the "message" folder and all of its contents

Note
I also removed the code in other files that contributes to the messages,
but it's part of larger changes to those so I'm putting them in a separate
commit.