Use postgres data for user class
Why
The user information is stored in the postgres database. It needs to be
used to initialize the user objects so the spotify tracks and response
sheets are specific to them

How
- Change the fields that are brought into the object. The spotify data
can be added at a later time once I have a better idea of how it should
be stored.
- Create an add_users function that will return the results of the query.
Right now, I just want to see what is returned so I can determine the best way to
use it for the user objects.